# backend/src/main/resources/contracts/autospec-backend-v1.openapi.yaml
openapi: 3.0.3
info:
  title: AutoSpec Backend API
  version: v1
  description: Spring Boot API contract for AutoSpec project orchestration, artifacts, exports, and observability.
servers:
  - url: /
security:
  - SessionToken: []
paths:
...
  /api/workflow-runs:
    post:
      summary: Start a versioned workflow with a frozen execution policy and server-trusted retrieval context.
      parameters:
        - $ref: '#/components/parameters/SessionToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowRunStartRequest'
      responses:
        '200':
          description: Started workflow run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRunResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/workflow-runs/{runId}:
    get:
      summary: Fetch a versioned workflow run and its budget consumption.
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Workflow run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRunResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/workflow-runs/{runId}/nodes:
    get:
      summary: List all node attempts for a workflow run.
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Workflow node attempts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowNodeRunResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/workflow-runs/{runId}/replay:
    post:
      summary: Create an immutable replay from the original or a selected workflow version.
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/SessionToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowReplayRequest'
      responses:
        '200':
          description: Replay workflow run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRunResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/workflow-runs/{runId}/nodes/page:
    get:
      summary: List workflow node runs by stable ascending ID cursor.
      description: New rows are returned only after the last ID from the previous page, avoiding offset drift.
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationCursor'
      responses:
        '200':
          description: Cursor page of workflow node runs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowNodeRunCursorPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    SessionToken:
      type: apiKey
      in: cookie
      name: AUTOSPEC_SESSION
  parameters:
    SessionToken:
      name: AUTOSPEC_SESSION
      in: cookie
      required: true
      schema:
        type: string
...
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    PaginationLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
    PaginationCursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the previous page; ordering is ascending by id.
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Missing or invalid session token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    Forbidden:
      description: Session user lacks the required project role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Project-scoped resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Operation conflicts with current durable workflow state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
  schemas:
...
    WorkflowExecutionPolicyRequest:
      type: object
      properties:
        qualityProfile:
          type: string
          enum: [FAST, BALANCED, DEEP]
        maxTokens:
          type: integer
          format: int64
          minimum: 1
        maxCost:
          type: number
          minimum: 0
          exclusiveMinimum: true
        maxModelCalls:
          type: integer
          minimum: 1
        maxWallTimeMs:
          type: integer
          format: int64
          minimum: 1000
    WorkflowRunStartRequest:
      type: object
      required: [projectId, workflowVersionId, input]
      properties:
        projectId:
          type: integer
          format: int64
        workflowVersionId:
          type: integer
          format: int64
        input:
          type: object
          description: Client requirement input. Retrieved sources are ignored and injected by the server from actor-accessible approved artifacts.
          additionalProperties: true
        idempotencyKey:
          type: string
          maxLength: 128
          nullable: true
        executionPolicy:
          $ref: '#/components/schemas/WorkflowExecutionPolicyRequest'
    WorkflowReplayRequest:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [ORIGINAL_SNAPSHOT, SELECTED_VERSION]
        selectedWorkflowVersionId:
          type: integer
          format: int64
          nullable: true
        idempotencyKey:
          type: string
          maxLength: 128
          nullable: true
    WorkflowRunResponse:
      type: object
      required: [id, projectId, operation, qualityProfile, consumedTokens, consumedCost, modelCallCount, status, startedAt]
      properties:
        id:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
        operation:
          type: string
        idempotencyKey:
          type: string
          nullable: true
        correlationId:
          type: string
          nullable: true
        workflowVersionId:
          type: integer
          format: int64
          nullable: true
        replayOfRunId:
          type: integer
          format: int64
          nullable: true
        reviewRound:
          type: integer
          nullable: true
        maxReviewRounds:
          type: integer
          nullable: true
        qualityProfile:
          type: string
          enum: [FAST, BALANCED, DEEP]
        maxTokens:
          type: integer
          format: int64
          nullable: true
        maxCost:
          type: number
          nullable: true
        maxModelCalls:
          type: integer
          nullable: true
        maxWallTimeMs:
          type: integer
          format: int64
          nullable: true
        consumedTokens:
          type: integer
          format: int64
        consumedCost:
          type: number
        modelCallCount:
          type: integer
        status:
          type: string
        responseStatus:
          type: string
          nullable: true
        responsePercent:
          type: integer
          nullable: true
        errorMessage:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
...
    WorkflowNodeRunCursorPage:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowNodeRunResponse'
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
    WorkflowNodeRunResponse:
      type: object
      required: [id, workflowRunId, nodeId, revision, attempt, executionId, status, handlerKey, handlerVersion]
      properties:
        id:
          type: integer
          format: int64
        workflowRunId:
          type: integer
          format: int64
        nodeId:
          type: string
        revision:
          type: integer
        attempt:
          type: integer
        executionId:
          type: string
        status:
          type: string
        handlerKey:
          type: string
        handlerVersion:
          type: string
        timeoutMs:
          type: integer
          nullable: true
        durationMs:
          type: integer
          nullable: true
        inputJson:
          type: string
          nullable: true
        outputJson:
          type: string
          nullable: true
        errorCode:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
        queuedAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        heartbeatAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        workerId:
          type: string
          nullable: true
...
    ApiErrorResponse:
      type: object
      required: [code, status, message, path, timestamp, fieldErrors, details]
      properties:
        code:
          type: string
        status:
          type: integer
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        details:
          type: object
          additionalProperties:
            type: string